+++ github/GithubIntegration.pyi
@@ -5,8 +5,6 @@ from github.InstallationAuthorization import InstallationAuthorization
 from github.PaginatedList import PaginatedList
 from github.Requester import Requester
 
-def create_jwt(integration_id: int = ..., private_key: str = ..., expiration: int = ..., issued_at: int = ...) -> str: ...
-
 class GithubIntegration:
     integration_id: Union[int, str] = ...
     private_key: str = ...
